[MNG-7667] Partial fix that "works for me"
This change hunts down m-install-p and m-deploy-p versions from
tested Maven (the one bound in default lifecycle) and resolves them
and their dependencies. This makes IT pass for me with both, non
updated Maven 3.9.0 and updated 3.9.0 re lifecycle plugins.

The ITs exectute this mojo in Bootstrap IT that executes with
pom that has packaging=pom, but the m-install-p and m-deploy-p
are still present. The "proper" fix would be to somehow grab the
versions of ALL plugins in all lifecycles burned in currently
tested Maven, but unsure how to achieve that.