[MNG-7667] Improve fix that cover m-p-p peculiarity
The m-p-p depends on m-p-tools-and and m-p-tools-beanshell as
runtime/optional, and the way DownloadMojo resolves every
artifacy causes that OptionalDependencySelector filter these
out even at "collection" stage (as depth >= 2).

This is NOT happening while m-p-p resolved as plugin, as there
m-p-p artifact POM is built and each dependency is being resolved
(hence all the "depth levels" are shifted for -1).

Doing this in DownloadMojo would tremendously slow down the process,
so adding a hack instead: when m-p-p is being resolved, we use
"hack" session that omits OptionalDependencySelector.

---

https://issues.apache.org/jira/browse/MNG-7667