[MNG-7706] Fix for ITs
The MNG-7706 deprecates ancient ArtifactRepository type use to
get access to local repository, and issues warning as for any
other deprecated Mojo parameters.

But alas, in ITs the MNG-5576 completely unrelated IT there is
an assertion to have WARNING-free log. This IT uses the
IT-plugins/IT-plugin-expression EvalMojo, that in turn
uses the deprecated `${localRepository}` parameter (but
does not use it). Result is, Maven 3.9.1 emits a WARNING
about use of deprecated parameter and the IT fails.

Further inspection shows, that while EvalMojo injects
ArtifactRepository for local repository, there is only
one IT that actually uses it, the MNG-4305, but even that
one is interested in basedir of the local repository only.
So to say, the use of deprecated ArtifactRepository is
not needed at all.

Fix:
* change EvalMojo to not expose in context the localRepository
  (w/ type ArtifactRepository), but a new expression `localRepositoryBasedir`
  only, that is injected in non-deprecated way.
* adjusted MNG-4305 to use new epxression instead to use object reflection in
  template to get basedir from ArtifactRepository
* This makes the originally failing MNG-5576 pass, as warning due EvalMojo
  is gone.

See
https://issues.apache.org/jira/browse/MNG-7706
https://github.com/apache/maven/pull/1009